Fix a bug when printing the operator
Only the original operand was printed, not the transformed one, if the operand
type transformed it. That is now fixed, so that the printed operand is the
same, but the operand-type gets added to the operator-string, which makes the
printed operation the original operation from the source-code, not the fully
decoded one.

Signed-off-by: Paul Schuetz <pa.schuetz@web.de>